:wrench: Fix int overflow crash problem

+++ core/src/main/java/com/ratchanon/CastleWarKnight/components/PointComponent.java
@@ -3,5 +3,5 @@ package com.ratchanon.CastleWarKnight.components;
 import com.badlogic.ashley.core.Component;
 
 public class PointComponent implements Component {
-    public int point;
+    public long point;
 }
